implemented part of the budget service tests

+++ internal/features/budgets/budget.go
@@ -1,8 +1,8 @@
 package budgets
 
 type Budget struct {
-	ID       int     `json:"_id"`
-	UserID   int     `json:"user_id"`
+	ID       int64   `json:"_id"`
+	UserID   int64   `json:"user_id"`
 	Name     string  `json:"name"`
 	Category string  `json:"category"`
 	Amount   float64 `json:"amount"`
